+++ docs/source/azureml.rst
@@ -6,25 +6,14 @@ AzureML module
 AzureML module from Recommenders utilities.
 
 AKS utils
-===================
+===============================
 
 .. automodule:: reco_utils.azureml.aks_utils
     :members:
 
 AzureML utils
-===================
+===============================
 
 .. automodule:: reco_utils.azureml.azureml_utils
     :members:
 
-SVD training utils
-===================
-
-.. automodule:: reco_utils.azureml.svd_training
-    :members:
-
-Wide and Deep training utils
-===================
-
-.. automodule:: reco_utils.azureml.wide_deep
-    :members:
