+++ docs/source/conf.py
@@ -18,7 +18,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath(os.path.join("..")))
+sys.path.insert(0, os.path.abspath(os.path.join("..", "..")))
 sys.setrecursionlimit(1500)
 
 from reco_utils import TITLE, VERSION, COPYRIGHT, AUTHOR
