+++ docs/source/recommender.rst
@@ -1,4 +1,95 @@
 .. _recommender:
 
 Recommender algorithms module
-**************************
+*******************************
+
+Recommender system algorithms from Recommenders utilities.
+
+
+DeepRec
+===============================
+
+.. automodule:: reco_utils.recommender.deeprec.models.base_model
+    :members:
+
+.. automodule:: reco_utils.recommender.deeprec.models.dkn
+    :members:
+
+.. automodule:: reco_utils.recommender.deeprec.models.xDeepFM
+    :members:
+
+.. automodule:: reco_utils.recommender.deeprec.IO.iterator
+    :members:
+
+.. automodule:: reco_utils.recommender.deeprec.IO.dkn_iterator
+    :members:
+
+.. automodule:: reco_utils.recommender.deeprec.deeprec_utils
+    :members:
+
+FastAI
+===============================
+
+.. automodule:: reco_utils.recommender.fastai.fastai_utils
+    :members:
+
+LightGBM
+===============================
+
+.. automodule:: reco_utils.recommender.lightgbm.lightgbm_utils
+    :members:
+
+NCF
+===============================
+
+.. automodule:: reco_utils.recommender.ncf.dataset
+    :members:
+
+.. automodule:: reco_utils.recommender.ncf.ncf_single_node
+    :members:
+    
+
+RBM
+===============================
+
+.. automodule:: reco_utils.recommender.rbm.rbm
+    :members:
+
+
+RLRMC
+===============================
+
+.. automodule:: reco_utils.recommender.rlrmc.RLRMCalgorithm
+    :members:
+
+.. automodule:: reco_utils.recommender.rlrmc.RLRMCdataset
+    :members:
+
+.. automodule:: reco_utils.recommender.rlrmc.conjugate_gradient_ms
+    :members:
+
+
+SAR
+===============================
+
+.. automodule:: reco_utils.recommender.sar.sar_singlenode
+    :members:
+
+Surprise
+===============================
+
+.. automodule:: reco_utils.recommender.surprise.surprise_utils
+    :members:
+
+
+Vowpal Wabbit
+===============================
+
+.. automodule:: reco_utils.recommender.vowpal_wabbit.vw
+    :members:
+
+Wide & Deep
+===============================
+
+.. automodule:: reco_utils.recommender.wide_deep.wide_deep_utils
+    :members:
