+++ docs/source/tuning.rst
@@ -1,4 +1,11 @@
 .. _tuning:
 
 Hyperparameter tuning module
-**************************
+*********************************
+Hyperparameter tuning module from Recommenders utilities.
+
+NNI utils
+===============================
+
+.. automodule:: reco_utils.tuning.nni.nni_utils
+    :members:
