+++ tests/unit/test_sweep.py
@@ -1,6 +1,6 @@
 import pytest
 
-from reco_utils.evaluation.parameter_sweep import generate_param_grid
+from reco_utils.tuning.parameter_sweep import generate_param_grid
 
 
 @pytest.fixture(scope="module")
